The fixtures move in: five kinds that furnish the hold
cargo.rs gains the room's surfaces: Mount { Ceiling, Floor, Wall },
Tag::Affix, and an affix rung in the placement ladder right after Cryo
— CeilingLamp and WallLamp (1x1), FloorLamp (1x2), Couch and Painting
(2x1), indices 16..=20 appended so every old save keeps its meaning.
Lamps are lit while stowed (lamp, lamp_lit, lit_adjacent join the sim's
public accessors, sharing the volatile rule's adjacency); rats fear the
light — no hop, boarding, or nibble beside a lit lamp, and a hold lit
wall to wall never boards a stowaway — while the couch tempts them into
Manhattan-step drift and a lazy, nibble-free nap that ordinary hops
eventually roll them out of. Insurance, not immunity.

barter.rs prices the new economy: Venus pays six for a Painting, Saturn
five for a repossessed Couch, the Hermitage six for comfort by its one
lit window, and the Umbra Market pays zero for every lamp — light is a
rival product, so they shelve seized lamps as cheap local produce,
snuffed at the till. A painting on the pads appraises +1 while a lit
lamp is stowed (the buyer sees it under your gallery lighting), proven
monotone by a second property test. The familiar discovery masks widen
u16 to u32 for the new kinds; the parser reads any hex width, so old
STV4 saves load untouched.

Thirteen new tests; both drag monkeys, the dial monotone property, and
the full 196-test suite green; frontends carry minimal placeholder arms
flagged for the fixture glyph pass.

+++ src/sim/save.rs
@@ -400,17 +400,19 @@ fn parse_parade(reader: &mut Reader<'_>) -> Result<(Option<u64>, Option<u64>), S
     Ok((parade_at, comet_visit))
 }
 
-/// The `familiar` line: one 4-hex kind bitmask per POI, in map order.
-fn parse_familiar(reader: &mut Reader<'_>) -> Result<[u16; POI_COUNT], SaveError> {
+/// The `familiar` line: one hex kind bitmask per POI, in map order. Written
+/// four digits wide for the old 16-kind masks; reads any width, so saves
+/// from before the fixture kinds (bits 16..) still load.
+fn parse_familiar(reader: &mut Reader<'_>) -> Result<[u32; POI_COUNT], SaveError> {
     let line = reader.next_line()?;
     let mut tokens = line.split_whitespace();
     if tokens.next() != Some("familiar") {
         return Err(reader.err());
     }
-    let mut familiar = [0_u16; POI_COUNT];
+    let mut familiar = [0_u32; POI_COUNT];
     for mask in &mut familiar {
         let token = tokens.next().ok_or_else(|| reader.err())?;
-        *mask = u16::from_str_radix(token, 16).map_err(|_| reader.err())?;
+        *mask = u32::from_str_radix(token, 16).map_err(|_| reader.err())?;
     }
     Ok(familiar)
 }
